backend/requirements.txt
Flask>=2.0
Flask-SocketIO
Flask-CORS
langchain
langchain-openai
openai
python-dotenv
eventlet
SQLAlchemy
langchain-community
